refactor: finish merge
Move sessionReducer into reducers.
