+++ src/reducers/index.js
@@ -1,18 +1,18 @@
 import { combineReducers } from 'redux'
 
-import account from '../reducers/account'
-import localChatReducer from '../reducers/localChatReducer'
-import IMReducer from '../reducers/IMReducer'
-import namesCoreReducer from '../reducers/nameReducer'
-import friendsReducer from '../reducers/friendsReducer'
-import sessionReducer from '../stores/sessionStore'
+import account from './account'
+import friendsReducer from './friendsReducer'
+import IMReducer from './IMReducer'
+import localChatReducer from './localChatReducer'
+import namesCoreReducer from './nameReducer'
+import sessionReducer from './sessionReducer'
 
 const rootReducer = combineReducers({
   account,
-  localChat: localChatReducer,
-  IMs: IMReducer,
-  names: namesCoreReducer,
   friends: friendsReducer,
+  IMs: IMReducer,
+  localChat: localChatReducer,
+  names: namesCoreReducer,
   session: sessionReducer
 })
